+++ src/main/java/com/gtc/tradinggateway/service/Withdraw.java
@@ -2,10 +2,12 @@ package com.gtc.tradinggateway.service;
 
 import com.gtc.tradinggateway.meta.TradingCurrency;
 
+import java.math.BigDecimal;
+
 /**
  * Created by Valentyn Berezin on 16.01.18.
  */
 public interface Withdraw extends ClientNamed {
 
-    void withdraw(TradingCurrency currency, double amount, String destination);
+    void withdraw(TradingCurrency currency, BigDecimal amount, String destination);
 }
