+++ src/main/java/com/gtc/tradinggateway/service/command/EsbCommandHandler.java
@@ -30,6 +30,7 @@ import org.springframework.jms.core.JmsTemplate;
 import org.springframework.stereotype.Service;
 
 import javax.validation.Valid;
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
@@ -81,7 +82,7 @@ public class EsbCommandHandler {
     public void getAllBalances(@Valid GetAllBalancesCommand command) {
         log.info("Request to create order {}", command);
         doExecute(accountTopic, command, accountOps, (handler, cmd) -> {
-            Map<TradingCurrency, Double> balances = handler.balances();
+            Map<TradingCurrency, BigDecimal> balances = handler.balances();
 
             log.info("Got balances {} for {} of {}", balances, cmd.getId(), cmd.getClientName());
             return GetAllBalancesResponse.builder()
@@ -100,8 +101,8 @@ public class EsbCommandHandler {
             OrderCreatedDto id = handler.create(
                     TradingCurrency.fromCode(cmd.getCurrencyFrom()),
                     TradingCurrency.fromCode(cmd.getCurrencyTo()),
-                    cmd.getAmount().doubleValue(),
-                    cmd.getPrice().doubleValue()
+                    cmd.getAmount(),
+                    cmd.getPrice()
             );
 
             log.info("Created {} for {} of {}", id, cmd.getId(), cmd.getClientName());
@@ -168,7 +169,7 @@ public class EsbCommandHandler {
         doExecute(withdrawTopic, command, withdrawOps, (handler, cmd) -> {
             handler.withdraw(
                     TradingCurrency.fromCode(cmd.getCurrency()),
-                    cmd.getAmount().doubleValue(),
+                    cmd.getAmount(),
                     cmd.getToDestination()
             );
 
@@ -177,7 +178,7 @@ public class EsbCommandHandler {
                     .clientName(cmd.getClientName())
                     .id(cmd.getId())
                     .currency(cmd.getCurrency())
-                    .amount(cmd.getAmount().doubleValue())
+                    .amount(cmd.getAmount())
                     .toDestination(cmd.getToDestination())
                     .build();
         });
